refactor(context): split File Evidence retrieval from its two adapters
Deepen the 1,215-line context/file-context.ts grab-bag into a File Evidence
retrieval module + two thin adapters (CONTEXT.md "File Evidence"):

- context/file-evidence.ts - the deep module: ~10 query primitives behind
  small `fileIds -> rows` interfaces + row types. No rendering, no composition.
  Unifies the two file resolvers into resolveFiles(paths, { fuzzyFallback })
  (hook hot path passes false to keep exact-only resolution).
- context/file-evidence-rank.ts - pure, DB-free signal extraction, ranking,
  and compaction. The unit-test surface for the logic that carries bugs.
- context/file-context-pack.ts - the CLI File Context Pack adapter
  (buildFileContextPack + renderers).
- hooks/file-context-hook.ts - gains buildFileContextHookEvidence (hook-only
  composition); injection/render logic unchanged.

Kill the dead slow loader: loadPriorFileSessions had two implementations - a
~3.5s/session per-row-subquery variant (pack) and the ~27ms batched lean
variant (hook). The slow one's only extra output (hands_free_ms) had no
consumer, so the pack now runs on the single lean loader (~130x faster).

Deliberate behavior changes (documented, both unread/cosmetic):
- pack prior_file_sessions[].hands_free_ms: real number -> null (no consumer;
  graph.tsx reads hands_free_ms from graph-explorer, not this pack).
- pack prior-session title now prefers the shipped PR title (the lean loader's
  existing hook behavior), instead of the first task-turn text.

Test surface: the one brittle SQL-substring composer mock splits into a
per-primitive pack stub + DB-free pure unit tests (signals, ranking,
compaction, numeric/duration/clip).

Gates: bun test (4731 pass / 0 fail) + bun run typecheck (clean).
